fix lint findings across the tree
Check json.Unmarshal/Decode/Put errors in tests, make ignored errors on
cleanup paths explicit with _ =, check the seed file Close error in
signer.WriteSeed, drop a dead jws assignment and an unneeded conversion,
and regroup imports so local packages form their own group (goimports
local-prefixes).

+++ tlogx/subtree_test.go
@@ -140,7 +140,7 @@ func hashTree(leaves []Hash, start, end uint64) Hash {
 }
 
 func bits63(n uint64) uint {
-	return uint(63 - leadingZeros(n))
+	return 63 - leadingZeros(n)
 }
 
 func leadingZeros(n uint64) uint {
